make `Iterator` chunkable
We want to parse the messages from byte streams represented by
`Iterator<Item=u8>`. For some types, like seed dictionaries, we'd like
to be able to get chunks of the byte stream. `itertools` provides such
an extension, but the resulting chunks do not implement
`ExactSizeIterator`. Therefore, we implement our own version of
it, which does implement `ExactSizeIterator`. Note that our iterator
is pretty limited because it forces the caller to fully consume all
the chunks in the order in which they are produced. However this is
an acceptable limitation for our use case.